Split affinity tests into their own class
